pyproject.toml
[tool.poetry]
name = "dramatron"
version = "0.1.0"
description = ""
authors = ["max brewer <deepfates@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.27.5"
python-dotenv = "^1.0.0"
rich = "^13.3.5"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
